Landlock ABI v6 support (scoped IPC)
This feature restricts the use of UNIX signals and dialing out to
Linux's Abstract UNIX Domain Sockets when the recipient process is in
a more privileged or unrelated Landlock domain.  When the restriction
is enabled, these forms of IPC are only permitted if the target
process is in the same or a less privileged (nested) Landlock domain.

Link: https://lore.kernel.org/all/cover.1725657727.git.fahimitahera@gmail.com/
Link: https://lore.kernel.org/all/cover.1725494372.git.fahimitahera@gmail.com/